feat : 노선에 등록되어 있는 역을 삭제할 경우 예외처리

+++ src/main/java/wooteco/subway/admin/service/StationService.java
@@ -4,15 +4,18 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.admin.domain.Station;
 import wooteco.subway.admin.dto.StationCreateRequest;
 import wooteco.subway.admin.dto.StationResponse;
+import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
 
 import java.util.List;
 
 @Service
 public class StationService {
+    private LineRepository lineRepository;
     private StationRepository stationRepository;
 
-    public StationService(StationRepository stationRepository) {
+    public StationService(LineRepository lineRepository, StationRepository stationRepository) {
+        this.lineRepository = lineRepository;
         this.stationRepository = stationRepository;
     }
 
@@ -35,6 +38,9 @@ public class StationService {
     }
 
     public void deleteById(Long id) {
+        if (lineRepository.countByStationId(id) > 0) {
+            throw new IllegalArgumentException("역이 노선에 존재합니다. 해당되는 노선에서 역을 모두 지우세요.");
+        }
         stationRepository.deleteById(id);
     }
 }
